Rebuild mfem with another hack to prevent VisIt compile errors. (#121)
the REGISTERED field in enum FlagMask collides with a Windows constant of the same name (when certain Windows headers are in the include chain, as happens when engine_ser is compiled). Renamed it to MREGISTERED everywhere it is used in mfem, and recompiled the library.
This fixes VisIt on Windows compile errors when mfem is used.